requirements.txt
httpx>=0.27
websockets>=12.0
pydantic>=2.7
pyyaml>=6.0
fastapi>=0.110
uvicorn[standard]>=0.29
jinja2>=3.1
pandas>=2.2
numpy>=1.26
pandas-ta>=0.3.14b0
loguru>=0.7
python-dotenv>=1.0
tenacity>=8.3
pytest>=8.2
pytest-asyncio>=0.23
freezegun>=1.5
